Late import `salt.client.ssh.client`
The reason behind this late import is that Salt's thin generation, if
Python 2 is available, tries to get the required dependencies, as such,
it tries to import `concurrent` provided by the `futures` package.
Having that import at the module scope will try to import `concurrent`,
if not available, the master will fail to start, even if the user is not
working with salt-ssh at all.

See also:
* https://github.com/saltstack/salt-bootstrap/pull/1466/checks?check_run_id=700200552
* https://github.com/saltstack/salt-bootstrap/pull/1466/checks?check_run_id=700200628